Correct extended error info for remote pids and ports
For BIFs that only accept a local pid, the extended error information would
state that the argument was not a pid. For example:

    (arne@seldon)1> RemotePid = rpc:call(kalle@seldon, erlang, whereis, [code_server]).
    <9325.50.0>
    (arne@seldon)2> process_info(RemotePid).
    ** exception error: bad argument
         in function  process_info/1
            called as process_info(<9325.50.0>)
            *** argument 1: not a pid

Correct the extended error information for those BIFs to report the argument as
"not a local pid". For example:

    (arne@seldon)1> RemotePid = rpc:call(kalle@seldon, erlang, whereis, [code_server]).
    <9325.50.0>
    (arne@seldon)2> process_info(RemotePid).
    ** exception error: bad argument
         in function  process_info/1
            called as process_info(<9325.50.0>)
    *** argument 1: not a local pid

Similarly, the extended error information for BIFs that operate on ports (such as
`port_control/3`) would state that a remote port was not a port.