stdlib: Implement `reason` key for error_info
Add so that the `reason` key if returned from format_error/2
is used to print the reason for the error when pretty printing.

Add so that also functions without arguments call format_error/2.

Both of these changes are there in order to integrate better
with the way that Elixir does error handling.