Handle when cover doesn't have any clause data
When cover runs against an erlang module that implements
functions for a NIF, cover doesn't have any clause data to
emit in its report. Currently, the code expects clauses always
exist for a module, but in this particular they don't and
it causes cover to crash when run in the context of rebar3.