diameter: fix set unodered flag in client (connect) mode
The semantics of switching a SCTP from ordered to unorderd are not
documented anywhere. The code as using assoc_id == 0, which mean
"all future associations". Adding a stream to the current association
is by strict reading not a new association, but an extension to the
current existing one. None the less, this work on the server
(listening) side.

It does not work on the client (connecting) side, at least on recent
Linux kernels. Using a value of 1 (SCTP_CURRENT_ASSOC) makes it work
on both sides.