ets:insert(list): Fix potential false-positive assert crash
The YCF extra context obtained by YCF_GET_EXTRA_CONTEXT() could change
between yields as the context is stack allocated before the first
yield. This could lead to a crash in debug builds as there is debug
code that checks that the data that is saved when yielding does not
contain pointers to the stack.