fix(kernel): do not bind sockets created from FDs
These sockets should be already bound for us, otherwise it is pointless
to pass them. In the end we are (almost) always able to create sockets
on our own (while we can not be able to bind it to some values). This
will check if socket is created form existing FD, and if so, it will
silently skip bind, assuming that it is already bound.

Fix #4680