GH-4396: exported function instead of fun to set up remote tracing
The function `dbg:n/1` used a local fun to set up a tracer on a remote
node. This works fine as long as the remote node is running exactly
the same version of Erlang/OTP but does not work at all
otherwise. This is fixed by exporting the relevant function
(`dbg:do_relay/2`) and by calling this function on the remote node to
set up remote tracing.