Correct order of optimization passes
81bc659e4cf0 switched the order of some optimization passes to
speed up compilation. Unfortunately, in rare circumstances that
would cause the `beam_ssa_bool` pass to do unsafe optimizations.

This commit eliminates the problem by running the `beam_ssa_bool` pass
before the `beam_ssa_share` pass (that is a partial revert of the
problematic commit). On my computer, that increases the compilation
time for the `String.Unicode` module in the Elixir standard library
with about 2 seconds (13 seconds instead of 11 seconds,
approximately).

Closes #4788.