Block enif_make_resource in destructor
Temporary fix to not accidentally call enif_make_resource from
the destructor.  The recource reference count is 0, hence
the call to the destructor, and then we can absolutely not call
enif_make_resource that will increase the reference count.

Then we have a NIF resource that is destructed, but not marked
as such.  Who knows what might happen.

The correct long term solution would be to ensure that the socket
is closed before the destructor is called, and to have an assertion
in the destructor.