Add app-file runtime_dependencies test and incorrect dependencies fixes
This commit adds a test case to check that app files for the
Erlang/OTP applications specifies correct versions for the
dependencies in the runtime_dependencies field. The test is designed
to work on the Erlang/OTP team's test servers as it assumes that all
released applications are installed in particular places and that all
app versions have been increased in the correct way (which is done by
scripts in the Erlang/OTP team's test infrastructure).

The test uses xref to try to find undefined function calls in all
Erlang/OTP applications. When running xref to test an application,
`xref`'s library path is set to point to folders containing beam-files
for the applications that are specified as the minimum correct
versions by the `runtime_dependencies` field in the application's app
file.

This commit also bumps the versions of several dependencies that the
test found to have incorrect versiosn so they instead get the first
version of the dependencies that makes the test pass.