compiler: Pretty-print types in BEAM SSA debug dumps
To help compiler developers, the compiler can be asked, during
compilation, to dump its internal BEAM SSA representation of the
compiled modules to a textual format. In it, type information is
printed as raw terms using the format specifier `~p` which for the
more complex types such as `#t_union{}` and `#t_tuple{}` become
rather hard to read.

This patch extends the BEAM SSA dumper with a pretty-printer for type
information. The BEAM SSA type information is similar to, but not the
same as, the representation used by Dialyzer so we can therefore not
use Dialyzer's type printer. The textual representation of BEAM SSA
types is designed to be as close to the Erlang Type Language as
possible.

Included in this patch is a test which checks that an ssa dump
contains the pretty printed types we expect for all but three output
forms. The exceptions are the bs_context and bs_matchable which never
seems to occur in function result types, and the interval form of
floats as the author is unable to construct a function which produces
such a result type (nor does such a type occur in any module compiled
by the diffable script).