ssl: Fix make deps
This commit fixes two issues. Firstly it makes it so that
the include files that we depend on is not taken from the
bootstrap system, which means that the prebuild should work.

Secondly it makes sure that if `$ERL_TOP` is something small
(like just `/OTP`), we do not do an overly zelous sed match.

Closes #4823