fix(kernel): init `socket` module on `kernel` start
`persistent_term` is cleaned on restart which causes the failures when
using `kernel.inet_backend = socket` as the `erl_epmd` module cannot
create connection. While this can slow down subsequent startups of
`kernel` application it ensures that the proper set of the protocols and
options is loaded to the `persistent_term` after restarts.

Close #4827