Add new lists:mapkeyfind/3 function to map finding
Stdlib's lists module has some functions for searching a list for a
tuple whose given element position compares equal to a given value, but
it lacks the same functionality for maps, i.e., a function for searching
a list for a map whose given key compares equal to the given value.

This new function allow us to find such a map in a list without the need
to create a fun to filter it every time a map search is required.

Usage:

1> lists:mapkeyfind(42, id, [#{id => 1, person => jane},
                             #{id => 42, answer => 42}]).
#{id => 42, answer => 42}

2> lists:mapkeyfind(not_found, id, [#{id => 1, person => jane},
                                    #{id => 42, answer => 42}]).
false

3> lists:mapkeyfind(not_found, id, []).
false