stdlib: Expand io docs to show that standard_io is group leader
This was previously not documented, but it is mentioned in a lot of
places that it is the group leader that all I/O is sent to by default.
So we can just as well make it explicit that standard_io is just an
alias for group_leader() and thus the target of standard_io is changed
when the group leader is changed.