Improve handling of exceptions from parse transforms
746fb27c38c7e1 (#4528) refactored handling of internal compiler
errors. An unintended consequence was that `exit` and `throw` exceptions
raised by parse transforms or core transforms would be reported as
internal compiler errors. As well as being confusing, the return value
from `compile:file/2` would be the atom `error` even when the `return_errors`
option was given. (The `return_errors` and `return_warnings` options are
intentionally ignored when the compiler has crashed in an unexpected way.)

Change the error handling so that exceptions in parse and core
transforms are reported as an error in the given transform. While at
it, also improve the formatting of the exception reasons and
stacktraces for transforms.