ssl: Make sure Trusted CA cert use latest PEM
The ssl application caches PEM files, that contain certificates and keys.
Trusted certs provided by the cacertfile option are also tracked in a special
cache to be able to share data between connections that uses the same input
and hence save memory. Make sure that this cache also gets invalidated when
the regular PEM caches is invalidated.