ssl: Extend certificate chain recreation logic
To be able to verify the certificate chain for more variants of
configuration, we now ignore expired trusted anchors at first.
This so that we can look for other possible valid issuers in a specific path,
that is a cross signed root certificate or another trusted certificate
present in the trust store but shadowed by the one that has expired.
If no alternative is found the connection will fail with a TLS certificate
expire ALERT.

This commit in addition to modifing path validation handling eliminates
unessessary code duplication from tls_handshake_1_3.

Closes #4877