Consolidate mandatory distribution flags across applications
The three applications that supports Erlang distribution (erts,
erl_interface, and jinterface), do not agree which the mandatory
distribution flags are.

For example, DFLAG_NEW_FLOATS (support for the "new" external format
for floats introduced in R11B-5) is considered mandatory in
erl_interface, but not in the Erlang emulator. jinterface does not
test the flag, but encodes all floats in the new external format.

Update erts, erl_interface, and jinterface to all use the same set of
mandatory distribution flags.

Also make the following distribution flags mandatory:

* DFLAG_EXPORT_PTR_TAG
* DFLAG_BIT_BINARIES
* DFLAG_MAP_TAG

Support for them in erl_interface and jinterface was added in OTP 22
or earlier. Making them mandatory allows us to remove ugly fallback
code.