kernel: Update read_file_info docs about symlinks
file:read_file_info follows all links, so it can never return
that a file is a symbolic link. However, read_link_info can of
course return a symbolic link and they use the same docs to
describe what can be returned. So we make it clear that link can
only be returned when read_link_info is called.

See #4917 for a bit more context.