RFC: diameter: allow `{error, Cause}` return from pick_peer/4
Currently, when pick_peer/4 can not pick a peer, it return false.
This is translated to a `{error, no_connection}` from diameter:call/4.

The are instance when the inability to choose a peer does not
mean that we don't have a connection to that peer. The main reason
is that we have somehow determined that all peers are overload or close
to beeing overload and want to moderate load.

This kind of load control can not be done the level of diameter:call/4
because at that level only the type of the call is know, not the
possible peers nor can it account for the failover to other peers.

By allowing errors returned from pick_peer/4, it can inform the
application that the request was actually dropped for other (e.g.
load) reasons.