erts: Fix use after free for ErtsTracerNif*
When doing `lookup_tracer_nif` the code assumes that
the ErtsTracerNif struct will not be deallocated after
the lookup. However, that is not true for unmanaged threads
as thread progress may pass after the lookup and before
the call to the tracer nif.

So, we make sure to delay thread progress before we do the
lookup to the ErtsTracerNif and undelay when done. As of the
writing of this commit there are 3 different trace flags that
can be triggered on a dirty scheduler:

 - garbage_collect
 - running
 - send