Introduce a scheme for reusing mandatory distribution flags
During distribution connect setup between two nodes, the nodes exchange
flags words encoding their capabilities. The connection setup is aborted
if one of the node lacks a capability that the other node requires.

In version 5 of the protocol (introduced in OTP R6), there are 32
possible capabilities (each encoded in a single bit). Since it seemed
that we would soon run out of capabilitiy bits, protocol version 6
(introduced in OTP 23) expanded the number of possible capability bits
to 64.

To avoid having to extend the number of capabilites yet again from 64
to 128 in some future release, this commit introduces a scheme to
allow reusing the bit numbers for capabilities that are now
mandatory. The scheme is described in lib/kernel/include/dist.hrl.

There are 10 mandatory capability bits as of OTP 25. This scheme will
allow reusing 9 of those capability bits as early as in OTP 27.