Optimize restart intensity calculation
add_restart/3 was a customized re-invention of lists:takewhile/2,
and as such even a misnomer, as it didn't add anything.
By putting the new restart on the list (in add_restart/1) before
passing it to add_restart/3 to be filtered caused an unnecessary filter
step, as the new restart would never be removed.
The inPeriod/3 function, used for checking if a restart was in or out
of period, did a calculation that could be replaced by a pre-calculated
constant value, as the Now and Period arguments were always the same
over a run of add_restart/3.

With this commit, the treshold time beyond which restarts are considered
as out of period is pre-calculated.
The filtering is done on the old list of restarts via lists:takewhile/2,
using the pre-calculated treshold.
Only then the new restart is added to the list of restarts.
This way, even the name add_restart starts to fit, and the inPeriod/3
function becomes obsolete.