[kernel] Improve inet:info/1
The info function did not handle a closed socket well.
A 'closed' socket is represented by a info map with *only*
the one or two states-attributes, with the value(s) set to
[closed].
Also if the function is called for a non-socket port, behave
as well as possible (no status or statistics).

OTP-17492

[kernel] More tweaking