jit: Provide the BeamModuleAssembler with an array of named labels
As the BEAM-file contains information about the name and arity of each
function, the information can be used to have asmjit pretty print the
labels starting the body of a function as name/arity.

This patch extends beam_load_prepare_emit() to retrieve all named
labels in the BEAM-file, sort them, and pass them on to the
assembler. Platform specific support will follow in subsequent
patches.