x86-jit: Use erlang-style <name>/<arity> for labels in asm dump
When running with `+JDdump true` the most verbose x86-assembly dump is
produced by the debug emulator. In it, labels present in the
beam-assembly are printed as `label_<index>`. This patch improves the
readability of the x86-assembly dumps by pretty-printing the label
marking the start of a function using its <name>/<arity> Erlang name.

Previously pretty-printing in the `label_<index>` format was only
available in the debug emulator. Readability and the ease with which a
user can cross-correlate x86-assembly and beam-assembly dumps is
significantly improved with pretty-printed labels, thus this patch
enables the functionality as soon as `+JDdump true` is enabled.