ssl: Correct cipher suite filtering
When TLS-1.3 is supported together with previous versions all cipher
suites is a concatenation of TLS-1.3 cipher suites and earlier
versions cipher suites.

Pre TLS-1.3 cipher suites are inclusive, that is are define
for all versions starting from when it was introduced and up to
TLS-1.2 with a few execprions.

Slight refactor to move all functions that handle explicit lists of
cipher suites to the tls_v1 module.

Document the fact that supplied cipher suites list that do not
comply with configuration and/or supported cryptolib will fallback to
appropriate default.

Document that pre TLS-1.3 PSK and SRP cipher suites need extra configuration
to work.

Closes #4978