arm-jit: Use erlang-style labels in asm dump
When running with `+JDdump true` the most verbose arm-assembly dump is
produced by the debug emulator. In it, labels present in the
beam-assembly are printed as `label_<index>`. This patch improves the
readability of the arm-assembly dumps by pretty-printing the label
marking the start of a function using the form
@Mod:Name/Arity-SequenceNumber.

Compared to the corresponding functionality on x86, this feature is
more complex to implement on Arm as with the trampolines used to
extend extend the range of branches, beam-level and Arm-assembly
labels no longer have one-to-one relationship.

Previously pretty-printing in the `label_<index>` format was only
available in the debug emulator. Readability and the ease with which a
user can cross-correlate arm-assembly and beam-assembly dumps is
significantly improved with pretty-printed labels, thus this patch
enables the functionality as soon as `+JDdump true` is enabled.