erts: Use AC_PROG_CXX to find C++ compiler
We previously used AC_CHECK_TOOLS() with a list of C++ compiler names
to find a C++ compiler. As of autconf 2.71 AC_CHECK_TOOLS() detects that
we try to find a C++ compiler and change mode to C++. It, however, fails
to restore to C mode when done and instead "restore" to C++ mode. This
causes the tests that follow to execute in C++ mode when C mode is
expected.

AC_PROG_CXX push and pop languages as expected. It also by default check
all C++ compiler names that we had in the list.

Since it works using autoconf 2.69, we only change this when using
autoconf 2.71.