erts: Add test case for erlang-style labels in asm dump
Check that pretty printing of named labels is working. We do that by
loading the test suite module in an emulator running with +JDdump true
and then checking that the produced jit_SUITE.asm contains a label for
each exported function. We also check that the label for the
non-exported function get_tmp_asm_files/0, which is used by this test,
is present.