Let a receiver of incoming NETCONF notifications be configured
Module ct_netconfc supports incoming notifications by way of the RFC
5277 create-subscription RPC, a process that calls create_subscription/*
receiving any subsequent notifications in its message queue. However,
create-subscription is no longer the only way to subscribe to
notifications: RFC 8639 added the establish-subscription RPC based on
RFC 7950, YANG 1.1. The possibility of defining new RPCs in YANG
invalidates the assumption that notifications are received only a
consequence of any fixed set of RPCs.

To be able to receive notifications regardless of how they originate,
add a 'receiver' option() to allow a destination for incoming
notifications (and possibly more in the future) to be specified when a
session is created; at session/*, open/*, or only_open/*. A destination
is anything that can be banged, as documented by erlang:send/2. Multiple
destinations can be specified and take precedence over the pid() of a
create_subscription/* caller. The behaviour is unchanged if no receiver
is specified.

Configuration of a receiver on connect/* should arguably either provide
a default or contribute additional receivers, but this is inconsistent
with option() capability, so a receiver on connect/* is simply ignored.
Otherwise, the option handling is a little inconsistent, with some
unused options being ignored and others being rejected. Possibly worth
fixing, but not in this commit.

There is no specific support for establish-subscription. Few of the
RPC-specific functions in ct_netconfc provide anything but simple
encode/decode, and there's little point in adding more functions when
the list of RPCs is potentially endless. Arbitrary RPCs can be sent with
send_rpc/*.