Fix code generation in guard for not (X =:= true)
In a guard, `not (X =:= true)` would incorrectly evaluate to `false`
for non-boolean values of `X`.

The is because of an unsafe transformation in the `v3_core` that
transforms `not (X =:= true)` to `X =:= false`. Unfortunately, that
transformation is only safe if `X` is known to be a boolean. (The
bug was introduced in OTP R13.)

Fixes #5007.