Eliminate redundant branches
`sys_core_fold` has a weak optimization that rewrites:

    case BoolExpr of
        true -> true
        false -> false
    end

to simply:

    BoolExpr

That optimization is weak because it will only handle a boolean
expression in a Core Erlang `case` expression, not other boolean
expressions resulting from, for example, guards.

Implement the optimization in a more effective way in the
`beam_ssa_opt` pass.