Do strength reduction of is_tagged_tuple and is_list
There is an instruction called `is_tagged_tuple` that is used for
matching a tuple whose first element is a literal atom. Here is
an example where this instruction will be used:

    f(X) ->
        case some_function(X) of
            {ok,Result} -> Result;
            Error -> Error
        end.

The following BEAM code will be generated:

        {allocate,0,1}.
        {call,1,{f,5}}.
        {test,is_tagged_tuple,{f,3},[{x,0},2,{atom,ok}]}.
        {get_tuple_element,{x,0},1,{x,0}}.
        {deallocate,0}.
        return.
      {label,3}.
        {deallocate,0}.
        return.

Better code can be generated if the compiler's type analysis has
figured out that the type of the return value of `some_function/1` is:

    {'ok', any()} | 'error'

If that is the case, since the return value of `some_function/1` is
either a tuple or the atom `error`, the `is_tagged_tuple` instruction
can be replaced with the less expensive (slightly faster and slightly
smaller) `is_tuple` instruction:

        {allocate,0,1}.
        {call,1,{f,5}}.
        {test,is_tuple,{f,3},[{x,0}]}.
        {get_tuple_element,{x,0},1,{x,0}}.
        {deallocate,0}.
        return.
      {label,3}.
        {deallocate,0}.
        return.

If the compiler has figured out that the type is:

    {'ok', any()} | {'error', any(), any()}

then `is_tagged_tuple` can be replaced with `test_arity`:

       {allocate,0,1}.
       {call,1,{f,5}}. % some_function/1
       {test,test_arity,{f,3},[{x,0},2]}.
       {get_tuple_element,{x,0},1,{x,0}}.
       {deallocate,0}.
       return.
     {label,3}.
       {deallocate,0}.
       return.

Unfortunately, if the type is:

    {'ok', any()} | {'error', any()}

which is not uncommon, there does not exist any alternative sequence
of instructions that would be guaranteed to be less expensive, so we
will keep `is_tagged_tuple`.

Also replace the `is_list/1` guard BIF with a less expensive instruction
when we have knowledge of the argument.