Do not attempt to start children of running apps
When optional applications were added in 79146711,
ensure_all_started/2 had to take explicit control
over the initialization cycle so available optional
applications were started properly. However, in
doing so, we lost one important optimization, which
was to not attempt to start children of a given
application if said application was already running.

This commit address this by exporting is_running/1
and skipping the work in case the given app is already
running. This change makes it so the time taken to
start a deep dependency tree goes from 58s to 4s.