sig_q_par_opt: Fix issue with signal sending from unmanaged threads
The signal queue optimization that allows several processes to enqueue
simultaneously without interfering with each other depends on that all
enqueuing threads are participating in the thread progress system to
ensure safe deallocation of its buffers array . Unfortunately, that
is not the case as, for example, NIFs can run on dirty schedulers,
and dirty schedulers are not participating in the thread progress
system.

This commit fixes this problem by introducing a reference counter for
the buffers array. The reference counter is only incremented by
threads that are not part of the thread progress system and prevents
deallocation of the buffers array while such threads access the buffers
array.