Fix distribution handshake challenge bug
When Multiple nodes would try to connect to each other, after having
each used erlang:set_cookie( OtherNode, OtherCookie ) the dist_util
handshake would fail.

This was in part due to auth:get_cookie(OtherNode) default behaviour
to return it's own Nodes cookie when a OtherNode was not in the cookie
ETS table.

The behaviour was as follows :

- NodeA uses NodeB's Cookie to establish connection:

  - NodeA challenge used NodeB's cookie, NodeB attempts to use NodeA's
    cookie but gets it's own cookie instead ( as NodeB did not use
    set_cookie for NodeA ), all cookies match in challenges successfully.

- NodeA & NodeB try to each establish connection to the other:

  - NodeA having set_cookie( NodeB, NodeBCookie )

  - NodeB having set_cookie( NodeA, NodeACookie )

  - Either Node attempts connection, i.e :

    - NodeA challenge used NodeB's cookie, NodeB used NodeA's cookie
      leading to mismatched challenges, failing to match.

The bug arose in dist_util.erl from mis-use/set of 'MyCookie' &
'HisrCookie'.

The behaviour now appropriately use sets the 'HisCookie' to the Other Node
and 'MyCookie' to it's own node using erlang:get_cookie().