Fix Our vs Other cookie handling
Correct the distribution handshake challenge bug the right way.

The intention of OurCookie vs. OtherCookie has been around since
before R7B, and was documented in R11B to be symmetrical, i.e that
if some node uses a different cookie then our node, then we need
to set that node's cookie both to connect to and to accept
connections from that node.  The same must then apply for
the other node, meaning that this relation is symmetrical.

So, for a configuration with different cookies to work,
all nodes has to have the cookies for all other nodes configured,
except for the nodes that use the same cookie as our node.

The same must apply for hidden nodes since it is a basic
symmetry property in the cookie system, and hidden nodes
just hides from e.g global to avoid a fully connected mesh.

The challenge exchange uses the other nodes's cookie for
outgoing messages, and our node's cookie when verifying
incoming messages.

A new internal function auth:get_cookies(Node) returns
{OurCookie,OtherCookie}, which for a node not having
a specific cookie set becomes {OurCookie,OurCookie}.

The code in dist_util.erl has been reverted to as it
has been since about R12B..R7B.., something, to use
OurCookie vs. OtherCookie in the challenge exchange.