[kernel|gen-tcp-socket] Handle buffered data when activating
If the socket is deactivated; active: once | true | N > 0 -> false
we do not cancel any select! Data that arrive during the phase when
we are in state 'recv' but are inactive is simply stored in the buffer.
Then if activated: active: false -> once | true | N > 0,
we need to check if there is something in our buffers, and maybe deliver
it to its owner. This should only occure if we are in state connected
(state 'recv' and in-active when data arrives => put data in buffer and
then enter state 'connected', since we are in-active).

OTP-17156