fix(logger): force atom format strings to strings
This is meant to prevent situations when user will fire messages using

    logger:notice(string, [$a]).

That would be accidentally treated the same as:

    logger:notice("a").

As earlier both produced message in form of `{string, "a"}`. Now it
simply cast format strings to never be an atom. This isn't breaking
change as message was defined to be `{io:format(), [term()]}`, which
mean that both forms should be treated identically.