[megaco|tcp-transport] TCP transport can now use inet_backend option
The option 'inet_backend' is now handled to make it possible
to specifically configure the inet-backend (inet or socket).

Also found what must be a very old bug in the tcp-transport
handling of transports (linkdb). It was supposed to be a list
but was initially just the atom 'undefined'. Result improper
list. Which may not have been an issue in practice if you only
used one transport...

OTP-17533