Update the doc CT test_spec grammar to match actual behavior
The grammar in the doc doesn't match actual behavior
Two problems:
Groups = GroupPath, GroupPath = [GroupName] leads to think that
[group1, group2] is the group path group1-group2 whereas it is
understood as an array of GroupSpec.
GroupSpec = [group, properties, GroupSpec] was not recognised, only
[group, properties, [GroupSpec]

An updated grammar is proposed, together with additional tests
to validate the pertinence of the new grammar.