erts: Fix FRAME_POINTER to be valid when stack adjusted
The frame pointer is not used by anything after hibernate, but some
ASSERTS assume that it is correct so we set it to the correct value
after the stack has been dropped when we hibernate.

Fault found by Max Federov and can be reproduced by running
hibernate_SUITE with the flags -emu_type debug +JPperf true.