Validate sockaddr tighter
Check that all fields in the sockaddr() value are part
of the address for the address family.

I have myself once too many used #{family =>inet, address => loopback}
and then gotten #{family => inet, addr => any} since the address
field is named 'addr' so 'address' is ignored and the default
for 'addr' is 'any' ({0,0,0,0}).