Fix inconsistency in defining ?FILE macro
This happens when the source_name is explicitly defined as an option to epp:parse_file/2. Before an include it got expanded to the actual file name, after an include it was defined as the source_name. With this change it will always be the source_name.

This come up when using --deterministic mode to compile causing the output to be less deterministic if you use the ?FILE macro before any includes.