erts: Further optimize off-heap traversal during minor GC
Only traverse first part of off-heap list that reside on the new-heap,
except for writable binaries that are now kept in a separate list.

Earlier during minor GC, traversing the off-heap list was continued
to the end over terms residing on the old-heap. This for two reasons:

1. The BIN_OLD_VHEAP value was calculated by adding the sizes
   of all ProcBin's on the old-heap.
2. All writable binaries are handled by shrinking (reallocating)
   binaries that have not been written to since last GC. Even binaries
   residing on the old-heap.

To optimize and only traverse the first part the off-heap list
that reside on the new-heap, we now

1. maintain a correct value of BIN_OLD_VHEAP so it does not have to be
   recalculated during GC.
2. keep all writable binaries in a separate list Process.wrt_bins
   that is traversed to the end during every GC.

The aim of this commit has been to maintain the overall GC semantics
and especially not alter the value of BIN_OLD_VHEAP or how binaries
have been reallocated at the end of a GC.