Make binary_to_term/1 more resilent against bad binaries
binary_to_term/1 could crash the runtime system when given a
corrupted binary containing the external representation of
maps.

This commit fixes two issues:

* There could multiple frees of the same memory area for
a corrupted term containing both large maps and small maps.

* Heap space would be allocated according to the number of elements
the map was supposed to contain, even if the binary was clearly too
short to contain that many elements.