Fix broken "fill-screen"-button in benchmark visualization HTML pages
The signal queue benchmark in parallel_messages_SUITE and the ETS
benchmark in ets_SUITE have benchmark result visualization HTML pages
with "fill-screen" buttons to make the graphs bigger. This button did
not work as intended before. When pressing the button for a graph, the
last graph got replaced with a bigger version and not the one over the
button. This is now fixed.