dialyzer: Refine the type for erlang:raise/3
Dialyzer and Typer would assume that `erlang:raise/3` always
returns `badarg`. That is a correct return value, but unless
`erlang:raise/3` is given incorrect arguments, it will raise an
exception.

The determined type would be strange for functions where
`erlang:raise/3` would always raise an exception. For example:

    foobar(A) ->
        try
            atom_to_list(A)
        catch
            C:Reason:Stk ->
                io:format("ERROR: ~p ~p\n~p\n", [C,Reason,Stk]),
                erlang:raise(C, Reason, Stk)
        end.

Dialyzer and Typer would think that the spec for `foobar/1` was:

    -spec foobar(_) -> 'badarg' | string().

Eliminate this issue by teaching Dialyzer to figure out the correct
return value or lack thereof for `erlang:raise/3` when the types
of its arguments are known. The spec for `foobar/1` will now be:

    -spec foobar(_) -> string().